fix(dbHandlerFactory): fix bug in test
Added a spy to check for a fake IndexedDbHandler (just a string). This was done because functions
inside the indexedDbHandler were called and unit tests have to just look at the tested unit.
Also there were bugs in creating the indexedDbHandler (at this moment).
Also removed an unnecessary and an at the moment unnecessary test in the dbHandlerFactoryTest.