test(dbHandler): add save and delete stream tests
Added tests to check if the save and delete streams work as intended.
Added MockDbHandler object to make it possible to build the correct streams with
the abstract DbHandler object.
Also added private subscriber properties to save the downstream subscriptions to the
delete and save upstreams. This has to be done to dispose these subscriptions in some tests.