refactor(dbHandler): remove getStoreKey() and getServerKey() functions
This was done because it is more efficient to assign them through a "keys" parameter in the dbHandler creation.
The "keys" property of the modelSchema is loaded on model creation anyway, so it is more efficient to use that property
than to search for the keys in the database schema.